requirements.txt
streamlit
openai
pandas
pdfplumber
PyMuPDF
reportlab
sendgrid
python-dotenv
pillow
